(1) Changed the way simulation steps are represented in Network3 (as doubles rather than longs) and (2) fixed a bug in network::update_concentrations() that was causing fixed species flags not to be respected.

+++ bng2/Network3/src/network.h
@@ -200,10 +200,14 @@ extern int   print_derivs_network(FILE* out);
 extern int   print_derivs_species_network(FILE* out);
 extern int   print_rates_network(FILE* out);
 extern int   print_derivs_byrxn_network(FILE* out);
-extern int   propagate_cvode_network(double* t, double delta_t, long int* n_steps, double* rtol, double* atol,
-									 int SOLVER_TYPE, long maxStep);
-extern int   propagate_euler_network(double* t, double delta_t, long int* n_steps, double h, long maxStep);
-extern int   propagate_rkcs_network (double* t, double delta_t, long int* n_steps, double tol, long maxStep);
+//extern int   propagate_cvode_network(double* t, double delta_t, long int* n_steps, double* rtol, double* atol,
+//									 int SOLVER_TYPE, double maxStep);
+//extern int   propagate_euler_network(double* t, double delta_t, long int* n_steps, double h, double maxStep);
+//extern int   propagate_rkcs_network (double* t, double delta_t, long int* n_steps, double tol, double maxStep);
+extern int   propagate_cvode_network(double* t, double delta_t, double* n_steps, double* rtol, double* atol,
+									 int SOLVER_TYPE, double maxStep);
+extern int   propagate_euler_network(double* t, double delta_t, double* n_steps, double h, double maxStep);
+extern int   propagate_rkcs_network (double* t, double delta_t, double* n_steps, double tol, double maxStep);
 
 extern FILE* init_print_concentrations_network(char* prefix, int append);
 extern int   print_concentrations_network(FILE* out, double t);
@@ -236,8 +240,9 @@ extern int   print_pca_network(FILE* out, double t);
 
 /* Gillespie Monte Carlo functions */
 extern int      init_gillespie_direct_network(int update_interval, int seed);
-extern int      gillespie_direct_network(double* t, double delta_t, double* C_avg, double* C_sig, long maxStep);
-extern long int gillespie_n_steps();
+extern int      gillespie_direct_network(double* t, double delta_t, double* C_avg, double* C_sig, double maxStep);
+//extern long int gillespie_n_steps();
+extern double gillespie_n_steps();
 extern double   gillespie_frac_species_active();
 extern double   gillespie_frac_rxns_active();
 extern void		delete_GSP_included();
